Fixed f2a.py and u2f.py to translate zitarev and added zitarev ugen to test.

+++ arco/src/o2audioio.h
@@ -218,7 +218,7 @@ public:
 
     // Keep track of minimum input buffer content
     int min_buffer_len;
-    int last_report_frame_count;
+    int64_t last_report_frame_count;
 
 
     const char *classname() { return O2audioio_name; }
